st: fix enlarge_buffer segment number check
This patch removes a bug in enlarge_buffer() that can make a read or
write fail under special conditions.

After changing TRY_DIRECT_IO to 0 and ST_MAX_SG to 32 in st_options.h, a
program that writes a first block of 128k and than a second bigger block
(e.g. 256k) fails. The second write returns errno EOVERFLOW, as
enlarge_buffer() checks the sg list and detects that it already is full.
As enlarge_buffer uses different page allocation orders depending on the
size of the buffer needed, the check does not make sense.

[joe: changelog]
Cc: Kai Makisara <kai.makisara@kolumbus.fi>
Signed-off-by: Bodo Stroesser <bstroesser@ts.fujitsu.com>
Signed-off-by: Joe Lawrence <joe.lawrence@stratus.com>