mpt2sas: move adapter watchdog code
Move the adapter watchdog and the code that starts and stops it into the source
file that includes the driver's PCI callback functions (in this case, the
driver's SCSI Host Layer code). The code shuffling keeps the watchdog code
private to its source file and removes a dependency between the driver's base
and SCSI code. This change also eliminates the need for a per-adapter
schedule_dead_ioc_flush_run callback pointer as the function and all of its
invocations are now in the same file.

Signed-off-by: Joe Lawrence <joe.lawrence@stratus.com>