[squash] kbuild: add klpclean to PHONY
Note: squash with ("kbuild: Support for Symbols.list creation")

The klpclean target doesn't generate a 'klpclean' file that could be
considered 'up to date', hence it is a PHONY target that should be run
every time it is requested.

A contrived example why this might be a problem: create a file of the
same name, then try to run 'make klpclean'.  Make sees the file and
considers it up to date, so the Symbols.list is never cleaned up.

  % touch klpclean
  % make klpclean
  make: 'klpclean' is up to date.
  % ls -l Symbols.list
  -rw-r--r--. 1 jolawren jolawren 2323179 Aug 29 16:55 Symbols.list

[joe: add klpclean to PHONY]

Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>