[squash] livepatch/klp-convert: fix klp-convert off-by-one sympos
Note: squash with ("livepatch: Add klp-convert tool")

There is a small bug in valid_sympos() when sympos > 0 and when the
the position is the last obj:sym in the set.

When there is a lone obj:sym name, we should expect its sympos=0,
however, when there are multiple obj:sym names found, its sympos
specifies the symbol indexed from 1 (not 0).

Split these two cases up to clarify the code. In addition to fixing the
above case, also strengthen the sympos=0 case by verifying that there is
indeed only a single obj:sym name found.

[joe: separate and fix valid_sympos() sympos=0 and !=0 checks]

Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>