[squash] klp-convert: make convert_rela() list-safe
Note: squash with ("livepatch: Add klp-convert tool")

The last thing convert_rela() does is to move the converted rela to
dedicated klp rela section by looking for symbol name matches in the
original section:

	/* Move the converted rela to klp rela section */
	list_for_each_entry_safe(r1, r2, &oldsec->relas, list) {
		if (r1->sym->name == r->sym->name) {
			list_del(&r1->list);
			list_add(&r1->list, &sec->relas);
		}
	}

convert_rela is called from the main() function while iterating over the
relas in each section using:

	 list_for_each_entry_safe(rela, tmprela, &sec->relas, list)

which is safe against list element removal for the current rela in
position, but *not* rela->list->next.  This situation could easily occur
if a relocation section contains back-to-back relas with the same symbol
name that need to be converted.

  % eu-readelf --relocs samples/livepatch/livepatch-annotated-sample.o
  Relocation section [ 2] '.rela.text' for section [ 1] '.text' at offset 0x98 contains 9 entries:
    Offset              Type            Value               Addend Name
    0x0000000000000001  X86_64_PC32     000000000000000000      -4 __fentry__
    0x0000000000000008  X86_64_PC32     000000000000000000      -4 saved_command_line     <<
    0x0000000000000017  X86_64_PC32     000000000000000000      -4 saved_command_line     <<
    0x000000000000001e  X86_64_32S      000000000000000000      +0 .rodata.str1.1
    0x0000000000000023  X86_64_PC32     000000000000000000      -4 seq_printf
    0x0000000000000031  X86_64_PC32     000000000000000000      -4 __fentry__
    0x0000000000000038  X86_64_32S      000000000000000000      +0 .data
    0x0000000000000051  X86_64_PC32     000000000000000000      -4 __fentry__
    0x000000000000003d  X86_64_PC32     000000000000000000      -4 klp_enable_patch

Modify convert_rela() to merely note the new klp rela section for
matching symbols.  This lets the main() function move each rela
individually as it safely iterates through the linked list.

Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>