livepatch: Enable the livepatch submodule
The final step when loading livepatch module for livepatching a module
is to actually enable the livepatch.

The steps are exactly the same as in kmp_module_comming().

Note that there is no need to call klp_cleanup_module_patches_limited()
in the error path. klp_add_object() modifies only the single matching
struct klp_patch.

Signed-off-by: Petr Mladek <pmladek@suse.com>