livepatch: Prevent infinite loop when loading livepatch module
Livepatch modules should get automatically removed when the patched
module gets removed. But the problem with forced flag has shown that
there might be situations where it did not work as expected.

The problem with forced flag is solved now. But there might be other
situations that are not known at the moment.

Be on the safe side and add a paranoid check for preventing an infinite
loop in klp_module_coming() callback. It might happen when it tries to load
a livepatch module that has already been loaded in the past and that
was not removed because of a bug somewhere.

POC NOTE: The main purpose of this patch is to show potential problems
	with the split livepatches. It is hard to say if the check is
	worth it or whether even more checks are needed.

Signed-off-by: Petr Mladek <pmladek@suse.com>