livepatch: match klp_resolve_symbols() with apply_relocate_add() syms
Now that late module patching is no longer supported and klp-relcations
are handled by the module loader, update klp_resolve_symbols() to
operate on symbol structures found in the load_info structure and not
the core_kallsyms copies.  This ensures that klp-symbol st_values are
properly set when apply_relocate_add() is called.

Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>