WIP - static keys klp-convert test case
Add another klp-convert test case that employs static keys.  The
livepatch module should reference static keys as defined elsewhere,
requiring klp-convert to resolve relocations in not only the text
section(s) but also the __jump_table as well.

Note: this test case is pretty shaky as we can crash the machine if we
      load the modules in opposite order: klp_test_keys then
      klp_test_keys_mod .

Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>