(squash) checkpatch: fix annotations
Note: squash with ("livepatch: Add klp-convert annotation helpers")

Fix the following checkpatch warnings:

  WARNING: __section(.klp.module_relocs.) is preferred over __attribute__((section(".klp.module_relocs.")))
  #28: FILE: include/linux/livepatch.h:243:
  +	__attribute__((__section__(".klp.module_relocs." #obj)))	\

  WARNING: space prohibited between function name and open parenthesis '('
  #29: FILE: include/linux/livepatch.h:244:
  +	__attribute__((aligned (4)))

  WARNING: __aligned(size) is preferred over __attribute__((aligned(size)))
  #29: FILE: include/linux/livepatch.h:244:
  +	__attribute__((aligned (4)))

Additional note: a simple s/__attribute__((aligned.../__aligned.../
replacement as suggested provokes the following cryptic assembler error
in the KLP_MODULE_RELOC macro:

  /tmp/ccPOST59.s: Assembler messages:
  /tmp/ccPOST59.s:232: Error: junk at end of line, first unrecognized character is `"'

which must be caused by a confusing stringification (or lack of) in the
preprocessor macros.  I avoided this by removing the prefix string
literal and concatenation altogether... not as pretty, but this form
appeases the assembler.

Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>